feat: better atc config
