+++ src/cmd/utils/type-check.ts
